requirements.txt
# =============================================================================
# REQUIREMENTS.TXT — Dependencias de NexoTime v2
# =============================================================================
# Instalar con: pip install -r requirements.txt

# ── Framework web ──
fastapi==0.115.0
uvicorn==0.30.6

# ── Base de datos ──
sqlalchemy==2.0.35
psycopg[binary]==3.2.4
# psycopg → driver moderno para PostgreSQL (v3, funciona mejor en Railway)

# ── Autenticación ──
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]==2.9.2

# ── Bot de Telegram ──
python-telegram-bot==21.6

# ── Scheduler (recordatorios programados) ──
apscheduler==3.10.4

# ── Zona horaria ──
pytz==2024.1

# ── Generación de gráficos ──
matplotlib==3.9.2
# matplotlib → para generar imágenes de estadísticas

# ── CORS (permitir peticiones desde la web) ──
# Ya viene incluido en FastAPI (starlette)

# ── Variables de entorno ──
python-dotenv==1.0.1
